% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_welfare.R
\name{est_welfare}
\alias{est_welfare}
\title{Estimate welfare differences for nested CES vs standard CES}
\usage{
est_welfare(x, x_diff, quantity, type, sigma_ces, sigma_nces, gamma_nces)
}
\arguments{
\item{x}{data.frame; data.frame of interest}

\item{x_diff}{data.frame; data with an added/deleted variety/variety group}

\item{quantity}{char; colname of quantity}

\item{type}{char; colname of type}

\item{sigma_ces}{num; elasticity of substitution for the Feenstra method}

\item{sigma_nces}{num; within product group elasticity of substitution (nested CES)}

\item{gamma_nces}{num; across product group elasticity of substitution (nested CES)}
}
\description{
Estimate welfare differences for nested CES vs standard CES
}
