% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ea_countries.R
\name{get_ea_countries}
\alias{get_ea_countries}
\title{Get Euro Area countries}
\usage{
get_ea_countries()
}
\value{
Returns a vector of all Euro Area countries.
}
\description{
Get Euro Area countries
}
