% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx_colors.R
\name{gx_colors}
\alias{gx_colors}
\title{My own color codes. https://www.sessions.edu/color-calculator/ can be used to find new colors.}
\usage{
gx_colors(theme = "light")
}
\arguments{
\item{theme}{char; "dark" or "light" theme colors}
}
\value{
Return my color palette. (HEX codes)
}
\description{
My own color codes. https://www.sessions.edu/color-calculator/ can be used to find new colors.
}
