% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx_theme.R
\name{gx_theme}
\alias{gx_theme}
\title{Initialize my ggplot theme.}
\usage{
gx_theme(
  plot = NULL,
  print_x = TRUE,
  print_y = TRUE,
  legend_title = FALSE,
  legend_position = "bottom",
  text_size = 12,
  theme = "light",
  text_colour = "#919397",
  legend_line_width = 1.2,
  rect_line_colours = "#919397",
  rect_line_width = 0.1,
  plot_background = "white",
  colours = gx_colors(),
  fill_colours = gx_colors(),
  legend_rows = 1,
  title = NULL,
  subtitle = NULL,
  x_title = NULL,
  y1_title = NULL,
  language = NULL
)
}
\arguments{
\item{plot}{ggplot object. By default Null and gx_theme can be assigned to plot but otherwise gx_theme will create it's own plot.}

\item{print_x}{Show x-axis. By default TRUE.}

\item{print_y}{Show y-axis. By default TRUE.}

\item{legend_title}{Whether to include a legend title or not default is FALSE.}

\item{legend_position}{Legend position. Default is bottom.}

\item{text_size}{Fontsize. Default is 14.}

\item{theme}{"dark" or "light" theme. Defaults to "light".}

\item{text_colour}{Text colour. Default is black.}

\item{legend_line_width}{Width of line in legend.}

\item{rect_line_colours}{Colour of horizontal help lines. Default is #7F7F7F.}

\item{rect_line_width}{Width of major grid lines. Different widths needed for various outputs; by default 0.0001.}

\item{colours}{Colours that should be used for aes color in ggplot. Default is gx_colors(). If both aes, fill and color, are used, by default both will start with yellow so one needs to choose the colours manually using on of the arguments (colour or fill_colour).}

\item{fill_colours}{Colours that should be used for aes fill in ggplot. Default is gx_colors(). If both aes, fill and color, are used, by default both will start with yellow so one needs to choose the colours manually using on of the arguments (colour or fill_colour).}

\item{legend_rows}{How many rows should the legend have. Does only work if plot is not null.}

\item{title}{Add plot title. Does only work if plot is not null.}

\item{subtitle}{add subtitle. Only works if plot is not nll.}

\item{x_title}{x-axis title. Does only work if plot is not null.}

\item{y1_title}{y1-axis title. Does only work if plot is not null.}

\item{language}{Format of Date and decimal point. By default NULL. In the default case system settings will be applied.}

\item{panel_background}{Background colour of plot area. Default is white.}
}
\value{
Return RBI theme to add to ggplot object.
}
\description{
Initialize my ggplot theme.
}
\examples{
library( "ggplot2" )
plot <- ggplot( data = iris,
                aes( x     = Sepal.Length,
                     y     = Sepal.Width,
                     color = Species ) ) +
          geom_line()
# change the main theme to RBI corporate design (colors are not changed)
\dontrun{
plot + gx_theme()
# change theme including colors
gx_theme( plot )
}
}
