% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_api.R
\name{mean_api}
\alias{mean_api}
\title{Calculates the mean of API gravity of different types of crude oil.}
\usage{
mean_api(x, weight = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{num; numeric vector}

\item{weight}{num; weighting vector}

\item{na.rm}{logical evaluating to TRUE or FALSE indicating whether NA values
should be stripped before the computation proceeds.}
}
\description{
Calculates the mean of API gravity of different types of crude oil.
}
\details{
As API gravity cannot be added linearly it is necessary to convert
   to specific gravity, then calculate means and only then convert back to
   API gravity.
   In practice, the difference between this method and taking means directly
   may not result into to large differences due to similarity of crude oils.
}
