% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oil_ngl_pipelines.R
\docType{data}
\name{oil_ngl_pipelines}
\alias{oil_ngl_pipelines}
\title{Global Oil Infrastructure Tracker, Global Energy Monitor, June 2022 release.}
\format{
## `oil_ngl_pipelines`
   A data frame with 92 rows and 35 columns:
}
\source{
<https://globalenergymonitor.org/projects/global-oil-infrastructure-tracker/download-data/>
}
\usage{
oil_ngl_pipelines
}
\description{
Contains information on pipeline connections around the world. Visit
Global Energy Monitor for a definition of the variables.
}
\keyword{datasets}
