% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_coir.R
\name{prepare_coir}
\alias{prepare_coir}
\title{Import and combine raw data files from the Crude Oil Import Register (data
   from 2013 to 2019)}
\usage{
prepare_coir()
}
\description{
Import and combine raw data files from the Crude Oil Import Register (data
   from 2013 to 2019)
}
\details{
No data transformations are applied. This function solely imports all
   Excel files and consolidates them into one data.frame. The ultimate
   dataset is included with the package installation and can be accessed
   directly. This function serves exclusively for replication purposes.
}
