% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_eurostat.R
\name{prepare_eurostat}
\alias{prepare_eurostat}
\title{Import and prepare Eurostat crude oil import data.}
\usage{
prepare_eurostat(only_eu = T)
}
\arguments{
\item{only_eu}{log; if FALSE column COMP_NR will be removed. Necessary if the
original dataset is used. Here we just use a subsample and consider only
imports into the European Union as a whole.}
}
\description{
Import and prepare Eurostat crude oil import data.
}
\details{
This function just renames variables, matches country names to country
   codes and creates some new variables from existing ones. No data
   transformations are applied.
}
