% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_quality_quantity.R
\name{merge_quality_quantity}
\alias{merge_quality_quantity}
\title{Merge Quality and Quantity data and adjust Russian misspecification}
\usage{
merge_quality_quantity(share_urals = 0.7)
}
\arguments{
\item{share_urals}{num; how much of Russian crude oil reported as Other
should be treated as Urals crude oil}
}
\description{
Merge Quality and Quantity data and adjust Russian misspecification
}
\details{
80% of imports from Russia are Urals"
Most imports from Russia are classified as "Other Russian Crude Oil". Thus,
no quality can be assigned. However, various EU Institutions state that 80\%
of imports from Russia are Urals. Also the more recent data on Eurostat
supports this fact. Therefore, we conservatively assume that 70\% of oil
imports from Russia are Urals. This parameter can be varied
for robustness checks. Lowering it to 50\% does not impair our results.
}
